Move some very low-level service classes to the base module
