+++ base/src/main/java/name/mlopatkin/andlogview/base/AppResources.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package name.mlopatkin.andlogview;
+package name.mlopatkin.andlogview.base;
 
 import com.google.common.io.ByteSource;
 import com.google.common.io.Resources;
@@ -33,6 +33,6 @@ public final class AppResources {
      * @throws IllegalArgumentException if the resource isn't available
      */
     public static ByteSource getResource(String resourcePath) {
-        return Resources.asByteSource(Resources.getResource(AppResources.class, resourcePath));
+        return Resources.asByteSource(Resources.getResource(AppResources.class, "../" + resourcePath));
     }
 }
